Modularize inventory utilities
