+++ server/routes/api/legacy.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Session
 
 from core.utils.db_session import get_db
 from core.utils.auth import get_current_user
-from server.routes.ui.devices import suggest_vlan_from_ip
+from modules.inventory.utils import suggest_vlan_from_ip
 from modules.network.models import VLAN
 from core.models.models import TablePreference
 import json
